refactor: add assets feature
